MAJ de l'application avec projet separé pour les classes utilisé dans le backend.
