+++ Calicot.Shared/Models/AuthenticateRequestUser.cs
@@ -1,0 +1,12 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Calicot.Shared.Models;
+
+public class AuthenticateRequestUser
+{
+    [Required]
+    public string UserName { get; set; } = default!;
+
+    [Required]
+    public string Password { get; set; } = default!;
+}
